8226384: Implement a better logic to switch between OpenGL and Metal pipeline
