+++ src/java.desktop/macosx/classes/sun/java2d/metal/MTLLayer.java
@@ -28,14 +28,11 @@ package sun.java2d.metal;
 import sun.java2d.NullSurfaceData;
 import sun.java2d.SurfaceData;
 import sun.lwawt.LWWindowPeer;
-import sun.lwawt.macosx.CFRetainedResource;
-
+import sun.lwawt.macosx.CFLayer;
 import java.awt.GraphicsConfiguration;
 import java.awt.Insets;
-import java.awt.Rectangle;
-import java.awt.Transparency;
 
-public class MTLLayer extends CFRetainedResource {
+public class MTLLayer extends CFLayer {
 
     private native long nativeCreateLayer();
     private static native void nativeSetScale(long layerPtr, double scale);
@@ -45,40 +42,13 @@ public class MTLLayer extends CFRetainedResource {
     private static native void validate(long layerPtr, MTLSurfaceData mtlsd);
     private static native void blitTexture(long layerPtr);
 
-    private LWWindowPeer peer;
     private int scale = 1;
-
-    private SurfaceData surfaceData; // represents intermediate buffer (texture)
 
     public MTLLayer(LWWindowPeer peer) {
         super(0, true);
 
         setPtr(nativeCreateLayer());
         this.peer = peer;
-    }
-
-    public long getPointer() {
-        return ptr;
-    }
-
-    public Rectangle getBounds() {
-        return peer.getBounds();
-    }
-
-    public GraphicsConfiguration getGraphicsConfiguration() {
-        return peer.getGraphicsConfiguration();
-    }
-
-    public boolean isOpaque() {
-        return !peer.isTranslucent();
-    }
-
-    public int getTransparency() {
-        return isOpaque() ? Transparency.OPAQUE : Transparency.TRANSLUCENT;
-    }
-
-    public Object getDestination() {
-        return peer.getTarget();
     }
 
     public SurfaceData replaceSurfaceData() {
@@ -102,10 +72,6 @@ public class MTLLayer extends CFRetainedResource {
             validate((MTLSurfaceData)surfaceData);
         }
 
-        return surfaceData;
-    }
-
-    public SurfaceData getSurfaceData() {
         return surfaceData;
     }
