+++ src/java.desktop/macosx/classes/sun/java2d/opengl/CGLLayer.java
@@ -26,56 +26,26 @@
 package sun.java2d.opengl;
 
 import java.awt.GraphicsConfiguration;
-import java.awt.Rectangle;
-import java.awt.Transparency;
-
 import sun.awt.CGraphicsConfig;
 import sun.java2d.NullSurfaceData;
-import sun.java2d.SurfaceData;
 import sun.lwawt.LWWindowPeer;
-import sun.lwawt.macosx.CFRetainedResource;
+import sun.java2d.SurfaceData;
+import sun.lwawt.macosx.CFLayer;
 
-public class CGLLayer extends CFRetainedResource {
+public class CGLLayer extends CFLayer {
 
     private native long nativeCreateLayer();
     private static native void nativeSetScale(long layerPtr, double scale);
     private static native void validate(long layerPtr, CGLSurfaceData cglsd);
     private static native void blitTexture(long layerPtr);
 
-    private LWWindowPeer peer;
     private int scale = 1;
-
-    private SurfaceData surfaceData; // represents intermediate buffer (texture)
 
     public CGLLayer(LWWindowPeer peer) {
         super(0, true);
 
         setPtr(nativeCreateLayer());
         this.peer = peer;
-    }
-
-    public long getPointer() {
-        return ptr;
-    }
-
-    public Rectangle getBounds() {
-        return peer.getBounds();
-    }
-
-    public GraphicsConfiguration getGraphicsConfiguration() {
-        return peer.getGraphicsConfiguration();
-    }
-
-    public boolean isOpaque() {
-        return !peer.isTranslucent();
-    }
-
-    public int getTransparency() {
-        return isOpaque() ? Transparency.OPAQUE : Transparency.TRANSLUCENT;
-    }
-
-    public Object getDestination() {
-        return peer.getTarget();
     }
 
     public SurfaceData replaceSurfaceData() {
@@ -95,10 +65,6 @@ public class CGLLayer extends CFRetainedResource {
             validate((CGLSurfaceData)surfaceData);
         }
 
-        return surfaceData;
-    }
-
-    public SurfaceData getSurfaceData() {
         return surfaceData;
     }
